requirements.txt
psycopg2
pandas
numpy
torch
scikit-learn
joblib
fire
fastapi
pydantic
uvicorn
requests
streamlit
streamlit-webrtc
av
opencv-python
plotly
scipy
aiosqlite
nats-py
streamlit-webrtc